requirements.txt
discord.py
apscheduler
strava.cz_api
